services/coach/requirements.txt
livekit-agents>=0.12
livekit-plugins-openai>=0.12
livekit-plugins-elevenlabs>=0.12
livekit-plugins-silero>=0.12
python-dotenv
fastapi
uvicorn
